refactor: move opcode table & util functions into separate module
